+++ temba/flows/models.py
@@ -1093,7 +1093,7 @@ class FlowSession(models.Model):
     uuid = models.UUIDField(unique=True)
     org = models.ForeignKey(Org, related_name="sessions", on_delete=models.PROTECT)
     contact = models.ForeignKey("contacts.Contact", on_delete=models.PROTECT, related_name="sessions")
-    status = models.CharField(max_length=1, choices=STATUS_CHOICES, null=True)
+    status = models.CharField(max_length=1, choices=STATUS_CHOICES)
 
     # the modality of this session
     session_type = models.CharField(max_length=1, choices=Flow.TYPE_CHOICES, default=Flow.TYPE_MESSAGE)
@@ -1118,7 +1118,7 @@ class FlowSession(models.Model):
     wait_started_on = models.DateTimeField(null=True)  # when it started waiting
     timeout_on = models.DateTimeField(null=True)  # when it should timeout (set by courier when last msg is sent)
     wait_expires_on = models.DateTimeField(null=True)  # when waiting run can be expired
-    wait_resume_on_expire = models.BooleanField(null=True)  # whether wait expiration can resume a parent run
+    wait_resume_on_expire = models.BooleanField()  # whether wait expiration can resume a parent run
 
     # the flow of the waiting run
     current_flow = models.ForeignKey("flows.Flow", related_name="sessions", null=True, on_delete=models.PROTECT)
@@ -1153,6 +1153,13 @@ class FlowSession(models.Model):
                 name="flows_session_voice_expires",
                 fields=("wait_expires_on",),
                 condition=Q(session_type=Flow.TYPE_VOICE, status="W", wait_expires_on__isnull=False),
+            ),
+        ]
+        constraints = [
+            # ensure that waiting sessions have a wait started and expires
+            models.CheckConstraint(
+                check=~Q(status="W") | Q(wait_started_on__isnull=False, wait_expires_on__isnull=False),
+                name="flows_session_waiting_has_started_and_expires",
             ),
         ]